chore: rename post-setup to pre-deploy
