+++ packages/deployment/src/__tests__/pre-pm2.spec.ts
@@ -32,8 +32,8 @@ describe('pre-pm2', () => {
           path: '/home/pm2/repository/issue-28',
           ref: 'origin/issue-28',
           repo: 'https://github.com/user/repository',
-          'post-setup':
-            "yarn install --frozen-lockfile && yarn deployment post-setup --branch-ref='origin/issue-28' --domain='my-domain.com' --database-url='postgresql://dbuser:dbpassword@127.0.0.1:5432/issue-28' && nginx -t",
+          'pre-deploy':
+            "yarn install --frozen-lockfile && yarn deployment pre-deploy --branch-ref='origin/issue-28' --domain='my-domain.com' --database-url='postgresql://dbuser:dbpassword@127.0.0.1:5432/issue-28' && nginx -t",
           'post-deploy':
             'PM2_HOME=./.pm2/ pm2 startOrRestart ./packages/deployment/dist/ecosystem.json && nginx -s reload',
         },
