--- packages/deployment/src/post-setup/index.ts
@@ -1,1 +1,0 @@
-export { postSetup } from './post-setup';
