+++ packages/deployment/src/pre-deploy/get-hosts.spec.ts
@@ -1,7 +1,7 @@
 import { getHosts } from './get-hosts';
-import type { PostSetupOptions } from './post-setup.types';
+import type { PreDeployOptions } from './pre-deploy.types';
 
-const defaultOptions: PostSetupOptions = {
+const defaultOptions: PreDeployOptions = {
   branchRef: 'origin/issue-n',
   domain: 'my-domain.com',
   databaseUrl: 'postgresql://dbuser:dbpassword@127.0.0.1:5432/issue-n',
@@ -10,7 +10,7 @@ const defaultOptions: PostSetupOptions = {
 describe('getHosts', () => {
   it('for main branch returns default hosts', () => {
     expect.hasAssertions();
-    const options: PostSetupOptions = { ...defaultOptions, branchRef: 'origin/main' };
+    const options: PreDeployOptions = { ...defaultOptions, branchRef: 'origin/main' };
 
     expect(getHosts(options)).toStrictEqual({
       components: 'components.my-domain.com',
@@ -21,7 +21,7 @@ describe('getHosts', () => {
 
   it('for other branches returns hosts with branchName as prefix', () => {
     expect.hasAssertions();
-    const options: PostSetupOptions = { ...defaultOptions, branchRef: 'origin/issue-n' };
+    const options: PreDeployOptions = { ...defaultOptions, branchRef: 'origin/issue-n' };
 
     expect(getHosts(options)).toStrictEqual({
       components: 'components-issue-n.my-domain.com',
