+++ packages/deployment/src/pre-deploy/get-ports.spec.ts
@@ -1,6 +1,6 @@
 import { findFreePorts } from 'find-free-ports';
 import { getPorts, DEFAULT_PORTS } from './get-ports';
-import type { PostSetupOptions } from './post-setup.types';
+import type { PreDeployOptions } from './pre-deploy.types';
 
 jest.mock<typeof import('find-free-ports')>('find-free-ports', () => {
   const actual = jest.requireActual('find-free-ports');
@@ -11,7 +11,7 @@ jest.mock<typeof import('find-free-ports')>('find-free-ports', () => {
   };
 });
 
-const defaultOptions: PostSetupOptions = {
+const defaultOptions: PreDeployOptions = {
   branchRef: 'origin/issue-n',
   domain: 'my-domain.com',
   databaseUrl: 'postgresql://dbuser:dbpassword@127.0.0.1:5432/issue-n',
@@ -20,7 +20,7 @@ const defaultOptions: PostSetupOptions = {
 describe('getPorts', () => {
   it('for main branch returns default ports', async () => {
     expect.hasAssertions();
-    const options: PostSetupOptions = { ...defaultOptions, branchRef: 'origin/main' };
+    const options: PreDeployOptions = { ...defaultOptions, branchRef: 'origin/main' };
 
     await expect(getPorts(options)).resolves.toStrictEqual(DEFAULT_PORTS);
   });
@@ -30,7 +30,7 @@ describe('getPorts', () => {
 
     const findFreePortsMock = findFreePorts as unknown as jest.Mock;
     findFreePortsMock.mockResolvedValueOnce([3005, 3006, 3007]);
-    const options: PostSetupOptions = { ...defaultOptions, branchRef: 'origin/issue-n' };
+    const options: PreDeployOptions = { ...defaultOptions, branchRef: 'origin/issue-n' };
 
     await expect(getPorts(options)).resolves.toStrictEqual({
       components: 3005,
@@ -44,7 +44,7 @@ describe('getPorts', () => {
 
     const findFreePortsMock = findFreePorts as unknown as jest.Mock;
     findFreePortsMock.mockResolvedValueOnce([3005, 3006, 3007]);
-    const options: PostSetupOptions = { ...defaultOptions, branchRef: 'origin/issue-n' };
+    const options: PreDeployOptions = { ...defaultOptions, branchRef: 'origin/issue-n' };
     await getPorts(options);
 
     expect(findFreePortsMock).toHaveBeenCalledWith(3, { startPort: 3005, endPort: 4000 });
