+++ packages/deployment/src/pre-deploy/index.ts
@@ -1,0 +1,1 @@
+export { preDeploy } from './pre-deploy';
