+++ packages/deployment/src/pre-deploy/pre-deploy.types.ts
@@ -4,7 +4,7 @@ export interface PortsMap {
   api: number;
 }
 
-export interface PostSetupOptions {
+export interface PreDeployOptions {
   /** The ref of the branch that is being published */
   branchRef: string;
   /** The domain name through which the server is accessible */
