+++ packages/deployment/src/pre-deploy/templates/ecosystem.json.ejs
