+++ packages/deployment/src/pre-deploy/templates/nginx.conf.ejs
