+++ packages/deployment/src/pre-pm2/pre-pm2.ts
@@ -28,7 +28,7 @@ function getDeploymentConfig(options: PrePm2Options): DeploymentConfig {
     path: getPath(options),
     ref: options.branchRef,
     repo: options.repository,
-    'post-setup': `yarn install --frozen-lockfile && yarn deployment post-setup --branch-ref='${options.branchRef}' --domain='${options.domain}' --database-url='${databaseUrl}' && nginx -t`,
+    'pre-deploy': `yarn install --frozen-lockfile && yarn deployment pre-deploy --branch-ref='${options.branchRef}' --domain='${options.domain}' --database-url='${databaseUrl}' && nginx -t`,
     'post-deploy':
       'PM2_HOME=./.pm2/ pm2 startOrRestart ./packages/deployment/dist/ecosystem.json && nginx -s reload',
   };
